Move dispatchTrackingRequests to a central location
